requirements.txt
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==3.1.0
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Moment==1.0.5

# Another extension
python-dotenv==0.21.0
SQLAlchemy==1.4.42
WTForms==3.0.1
email_validator